refactor: entity -> converter 네이밍 변경
