--- src/main/java/repository/entity/TurnEntity.java
@@ -1,13 +1,0 @@
-package repository.entity;
-
-import janggi.piece.Team;
-
-public record TurnEntity(
-        String turn
-) {
-
-    public static TurnEntity toEntity(Team team) {
-        if(team==Team.CHO)return new TurnEntity("CHO");
-        return new TurnEntity("HAN");
-    }
-}
